+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on "Keep a Changelog".  This project adheres to Semantic Versioning.
 
 
+## [2.2.1] - 2022-10-13
+- Updated to work in Python 3 too
+
+
 ## [2.2.0] - 2022-04-21
 - Added new classes: RabbitMQBase and RabbitMQApi